requirements.txt
autopep8==1.4.3
boto3==1.9.180
botocore==1.12.180
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
cssmin==0.2.0
docutils==0.14
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
Flask-Login==0.4.1
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-vision==0.37.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsmin==2.2.2
MarkupSafe==1.1.0
numpy==1.16.4
opencv-python==4.1.0.25
peewee==3.8.2
peewee-db-evolve==3.7.0
protobuf==3.8.0
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
webassets==0.12.1
Werkzeug==0.14.1
wincertstore==0.2
